api/server/requirements.txt
aniso8601==7.0.0
asgiref==3.3.4
Django==3.2.3
django-filter==2.4.0
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
numpy==1.20.3
pandas==1.2.4
promise==2.3
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
Rx==1.6.1
singledispatch==3.6.2
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
typing-extensions==3.10.0.0
